migrated mredis (non-async)
